pyproject.toml
[project]
name = "citation-agent"
version = "0.1.0"
description = "CLI helper that uses OpenAI agents to extract and verify legal citations."
authors = [
  { name = "Your Team", email = "support@example.com" }
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "openai-agents>=0.5.0",
  "typer>=0.15.1",
  "pydantic>=2.9.0",
  "rich>=13.9.4"
]

[project.optional-dependencies]
pdf = ["pypdf>=5.1.0"]
docx = ["python-docx>=1.1.2"]

[project.scripts]
citation-agent = "citation_agent.cli:app"

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"
